requirements.txt
absl-py==0.12.0
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
autocrop==1.1.1
autopep8==1.5.7
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bleach==4.1.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
easydict==1.6
entrypoints==0.3
filelock==3.3.0
flatbuffers==1.12
gast==0.3.3
gdown==4.0.2
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.5.0
importlib-resources==5.2.2
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
joblib==0.10.3
jsonschema==3.2.0
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Keras==2.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.4
numpy==1.18.5
oauthlib==3.1.1
olefile==0.46
opencv-python==3.4.1.15
opt-einsum==3.3.0
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.2
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pydeck==0.6.2
Pygments==2.10.0
pymediainfo==5.0.3
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.3
PyWavelets==1.1.1
pywin32==302
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scipy==1.4.1
Send2Trash==1.8.0
six==1.15.0
sk-video==1.1.10
smmap==4.0.0
streamlit==1.0.0
tensorboard==1.14.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==1.14.0
tensorflow-addons==0.13.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
Theano==1.0.5
tifffile==2020.9.3
toml==0.10.2
toolz==0.11.1
torch==1.8.1
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typeguard==2.13.0
typing-extensions==3.7.4.3
tzdata==2021.2.post0
tzlocal==3.0
urllib3==1.26.5
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.1
